requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
appnope==0.1.0
astroid==1.6.1
async-timeout==2.0.0
asyncio==3.4.3
autopep8==1.3.3
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2018.1.18
chardet==3.0.4
cleverwrap==0.2.3.6
ConfigArgParse==0.9.3
decorator==4.0.11
Django==1.8
emoji==0.4.5
entrypoints==0.2.3
fudge==1.1.1
futures==3.1.1
google-api-python-client==1.6.5
goslate==1.5.1
hangups==0.2.10.2
hangups-urwid==1.2.2.dev0
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
inflect==0.2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.3
jaraco.itertools==2.1
jedi==0.10.2
Jinja2==2.9.6
jsonrpclib-pelix==0.3.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
MechanicalSoup==0.4.0
mistune==0.7.4
more-itertools==4.1.0
multidict==4.1.0
nbconvert==5.2.1
nbformat==4.3.0
nltk==3.2.5
notebook==5.0.0
oauth2client==4.1.2
oauthlib==2.0.6
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
purplex==0.2.4
pushbullet.py==0.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.8.2
pyslack-real==0.6.0
python-dateutil==2.6.0
python-magic==0.4.15
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
readlike==0.1.2
ReParser==1.4.3
requests==2.6.0
requests-oauthlib==0.8.0
rsa==3.4.2
selenium==3.9.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
slackbot==0.5.1
slackclient==1.1.2
slacker==0.9.60
soundcloud==0.5.0
spotipy==2.4.4
telepot==7.1
terminado==0.6
testpath==0.3.1
textblob==0.15.1
tornado==4.5.1
traitlets==4.3.2
TwitterAPI==2.4.8
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.46.0
widgetsnbextension==2.0.0
wikipedia==1.4.0
wolframalpha==3.0
wrapt==1.10.11
xmltodict==0.11.0
